exportercustomizer/gcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opentelemetry-instrumentation-exporter-customizer-component"
dynamic = ["version"]
description = "OpenTelemetry OTLP exporter customizer"
readme = "README.rst"
license = "Apache-2.0"
requires-python = ">=3.8"
authors = [
  { name = "OpenTelemetry Authors", email = "cncf-opentelemetry-contributors@lists.cncf.io" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "google-auth ~= 2.38",
  "grpcio ~= 1.7",
  "opentelemetry-exporter-otlp-proto-grpc ~= 1.3",
  "opentelemetry-exporter-otlp-customizer ~= 1.3",
]

[project.entry-points.opentelemetry_otlp_exporter_customizer]
google_otlp_customizer = "opentelemetry.auth.gcp:GoogleOTLPExporterCustomizer"

[tool.hatch.version]
path = "src/opentelemetry/auth/gcp/version.py"

[tool.hatch.build.targets.sdist]
include = [
  "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/opentelemetry"]